requirements.txt
flask
pillow
torch
torchvision
torchaudio
transformers
diffusers
accelerate
safetensors
mediapipe
